+++ src/main/java/chess/domain/position/Row.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.domain.position.exception.InvalidRowException;
+import chess.exception.InvalidRowException;
 
 import java.util.Arrays;
