+++ src/main/java/chess/exception/InvalidColumnException.java
@@ -1,4 +1,4 @@
-package chess.domain.position.exception;
+package chess.exception;
 
 public class InvalidColumnException extends PositionException {
     public static final String INVALID_COLUMN_NAME_MESSAGE = "유효하지 않은 열입니다.";
