+++ src/main/java/chess/exception/InvalidRowException.java
@@ -1,4 +1,4 @@
-package chess.domain.position.exception;
+package chess.exception;
 
 public class InvalidRowException extends PositionException {
     public static final String INVALID_ROW_NAME_MESSAGE = "유효하지 않은 행입니다.";
