+++ src/test/java/chess/domain/position/ColumnTest.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.domain.position.exception.InvalidColumnException;
+import chess.exception.InvalidColumnException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
