+++ src/test/java/chess/domain/position/RowTest.java
@@ -1,6 +1,6 @@
 package chess.domain.position;
 
-import chess.domain.position.exception.InvalidRowException;
+import chess.exception.InvalidRowException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
